Changed the modification date of the Release files.
It isn't saved in GIT!
Think to check the modification date of the Release files before packaging.
If the date is set to a too near date, these files won't be updated when
running apt-get update, and that will lead to checksum mismatchs !
A good practice would be to set this date to the release date of
the version.