Applied work-around to debian/ubiquity.templates
Some issue with debconf results in a failure to install ubiquity's deb.
Unknown what causes it, but a work-around is to remove all the comments
from debian/ubiquity.templates.

Debconf still complains, but installation now works. There are dozens of
warnings like this which still need to be fixed.

debconf: Unknown template field '_description', in stanza #3 of
/var/lib/dpkg/info/ubiquity.templates